beers seeder data functions complete

+++ backend/db/migrations/20211011191555-create-beer.js
@@ -16,7 +16,7 @@ module.exports = {
         type: Sequelize.STRING(750)
       },
       abv: {
-        type: Sequelize.INTEGER
+        type: Sequelize.NUMERIC(2, 1)
       },
       ibu: {
         type: Sequelize.INTEGER
